添加 后台管理全选和删除多个

+++ personBlog/src/cn/edu/tzc/blog/controller/admin/article/AdminArticleDeleteController.java
@@ -43,7 +43,13 @@ public class AdminArticleDeleteController extends HttpServlet {
 			if("" == ids || ids == null) {
 				message = articleService.DeleteAllArticle(user.getId(),path);
 			}else {
-				message = articleService.DeleteArticle(Integer.parseInt(ids),path);
+				if(ids.contains(",")) {
+					String[] idList = ids.split(",");
+					message = articleService.delCheck(idList, filePath);
+				}
+				else {
+					message = articleService.DeleteArticle(Integer.parseInt(ids),path);
+				}
 			}
 			String url = request.getContextPath()+"/admin/article";
 			pw.write("<html><body><script language='javascript'>alert("+message+");window.location.href='\"+url+\"';</script></body></html>");
